Remove dic_users
This dictionary hard-codes a list of users.  We don't use it at the
moment as the user id is populated from the system settings.  I'm
removing it so that we have the most minimal setup possible. We are
likely to need a table for user-specific metadata and settings in the
future, so it can be added in then.