Reference dic_superficial_code table directly
In order to have easy filtering in QGIS, we need a separate table with
the categories of different superficial landforms.  These correspond to
different groups within the code dictionary.  Previously the foreign key
dependency was on the category table, but actually, this table is just
for helping so the relationship has been updated to point at the code
table.  This makes it consistent with the structural measure format.