Use x and y instead of lat, lon in names
Projected coordinates in the UK are often called easting and northing,
which would make it an appropriate substitute for longitude and
latitude.  However, in some countries X increases to the west, or even
to the south, so it is best if we don't make any assumptions.  GDAL
/ Proj will understand these local variations and they use X, Y, so we
should report what they say.