Change message asking for project creation
The tone of the old message was a bit aggressive.  Also, if you create
a project and then try to add test data you get an error as the test
data input script specifies fid of 1, which has been used by the newly
created project.  This message highlights that it is one or the other.