Add label and category to _lnk_field_project
This commit adds two extra columns to lnk_field_project so that users
can see more information about the rows that they are editing than just
the opaque codes.  A trigger populates these values when a new code is
added.  The layer properties were updated to ensure that all available
rows appear in the dropdown and to make the extra columns read-only.