Update QGIS to 3.40 and add pillow dependency
This will only successfully build an environment if you use the libmamba
solver for conda.  It is also necessary to unpin the SQLite version as
it cannot be resolved with QGIS 3.40.  Once built, the tests fail with a SQLite error.  The next step is to work out how to fix this.

Only use conda-forge repository for packages

The SQLite issue that we had was because the version of sqlite3 was much
older than libsqlite3.  The former came from conda, the latter from
conda-forge.  Taking everything from conda-forge allowed a better match
between versions.

In general, conda/main is preferable as the packages have had more
curation.  However, in this case, a working environment is the priority.

Move conda-installable packages to conda section

Remove unused cx_Oracle

Update environment.yml to QGIS 3.40