Use project.setDirty() after map face note setting change
The extended map face notes setting changes the project and requires the
project to be saved for the changes to persist.  By using setDirty(),
this marks the title bar with a * so that the user is aware of unsaved
changes. It also asks the user to save changes on closing the project.