Remove architecture specification
The architecture on developer laptops is different, resulting in
different microarchitecture packages being selected by conda.  These are
two examples:

_x86_64-microarch-level=4=2_icelake
_x86_64-microarch-level=3=2_skylake

It is possible to create a new environment without specifying a pinned
architecture and let conda sort it out.  However, it is necessary to
have the "archspec" conda package available in the "base" environment
when you run the "conda env create" command.